refactor: Split up AddModelDialog.
